fix: avoid sign-conversion in GeometryHierarchyMap::find
Dropping -isystem for in-tree Acts headers exposes this header to
traccc's -Werror -Wsign-conversion, where the signed iterator distance
used to index m_ids/m_masks becomes a hard error.

Hold the index as std::size_t and convert once at the std::next call.
The 'while (0 < i) { --i; ... }' guard keeps the unsigned decrement safe.

+++ Core/include/Acts/Geometry/GeometryHierarchyMap.hpp
@@ -12,6 +12,7 @@
 
 #include <algorithm>
 #include <cassert>
+#include <cstddef>
 #include <initializer_list>
 #include <iterator>
 #include <stdexcept>
@@ -292,7 +293,7 @@ inline auto GeometryHierarchyMap<value_t>::find(
   // be sorted before the requested id. searching for the first element
   // after the requested ensures that we include the full hierarchy.
   const auto it = std::ranges::upper_bound(m_ids, id.value());
-  auto i = std::ranges::distance(m_ids.begin(), it);
+  auto i = static_cast<std::size_t>(std::ranges::distance(m_ids.begin(), it));
 
   // now go up the hierarchy to find the first matching element.
   // example: the container stores four identifiers
@@ -326,7 +327,7 @@ inline auto GeometryHierarchyMap<value_t>::find(
     // progresses from more specific to less specific elements. the first
     // match is automatically the appropriate one.
     if (equalWithinMask(id.value(), m_ids[i], m_masks[i])) {
-      return std::next(begin(), i);
+      return std::next(begin(), static_cast<std::ptrdiff_t>(i));
     }
   }
